Reworking of Workout Creation
Date picker and POST request now actioned on main 'workout.js' file. New button added for each workout, linking to the 'create.js' file, which will now be used solely for added activities.

This means activities can now be added after the initial creation of the model, but is separate to the details page to avoid over-encumbering the API